tests passing for registration, login, pantry routes, and loginform, addingredient, individualingredient, header components
