+++ package.json
@@ -17,6 +17,7 @@
     "react-modal": "^3.11.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.3.0",
+    "react-test-renderer": "^16.12.0",
     "styled-components": "^5.0.1",
     "styled-react-modal": "^2.0.0"
   },
