+++ src/Components/Pantry/AddIngredient.test.js
@@ -1,20 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import DashboardRoute from './DashboardRoute';
+import AddIngredient from './AddIngredient';
 import { BrowserRouter } from 'react-router-dom';
 import renderer from "react-test-renderer";
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
   ReactDOM.render(<BrowserRouter>
-    <DashboardRoute />
+    <AddIngredient />
   </BrowserRouter>, div);
   ReactDOM.unmountComponentAtNode(div);
 });
 
 it('renders the UI as expected', () => {
   const tree = renderer.create(<BrowserRouter>
-    <DashboardRoute />
+    <AddIngredient />
   </BrowserRouter>)
     .toJSON();
   expect(tree).toMatchSnapshot();
