+++ src/Components/Pantry/IndividualIngredient.js
@@ -1,167 +1,169 @@
-// import React, { Component } from "react";
-// import TokenService from '../../Helpers/Token'
-// import config from '../../config';
-// import "./Pantry.css";
+import React, { Component, useState } from "react";
+import TokenService from '../../Helpers/Token'
+import config from '../../config';
+import styled from "styled-components";
+import Modal, { ModalProvider, BaseModalBackground } from "styled-react-modal";
+import "../../index.css";
+import { render } from "@testing-library/react";
 
 
+export default class FancyModalButton extends Component {
+  constructor(props) {
+    super(props)
 
-// class IndividualIngredient extends Component {
-//   constructor(props) {
-//     super(props);
-
-//     this.state = {
-//       ingredient: [],
-//       updateIngredient: false
-//     };
-//   }
-
-//   componentDidMount() {
-//     this.setIngredient();
-//   }
-
-//   setIngredient = () => {
-//     const originalIngredient = this.props.location.state;
-//     console.log("original ingredient", this.props.location.state);
-
-//     this.setState({
-//       ingredient: originalIngredient
-//     });
-//   };
-
-//   handleSubmit = e => {
-//     e.preventDefault();
-//     let ingredientId = this.props.match.params.ingredientId;
-//     const url = `${config.API_ENDPOINT}/pantry/${ingredientId}`;
-//     const authToken = TokenService.getAuthToken();
-//     let { ingredient_name, in_stock, notes } = e.target;
-
-//     let updatedIngredient = {
-//       id: ingredientId,
-//       ingredient_name:
-//         ingredient_name.value || this.props.location.state.ingredient_name,
-//       in_stock: in_stock.value || this.props.location.state.in_stock,
-//       notes: notes.value || this.props.location.state.notes
-//     };
-//     console.log(
-//       "updated ingredient to be sent to server is",
-//       updatedIngredient
-//     );
-//     fetch(url, {
-//       method: "PATCH",
-//       headers: {
-//         "content-type": "application/json",
-//         "authorization": `Bearer ${authToken}`
-//       },
-//       body: JSON.stringify(updatedIngredient)
-//     })
-//       .then(res => {
-//         if (!res.ok) return res.json().then(error => Promise.reject(error));
-//       })
-//       .then(data => {
-//         console.log("patch data is", data);
-//         this.props.history.push("/pantry");
-//       })
-//       .catch(error => {
-//         console.error(error);
-//       });
-//   };
-
-//   handleGoBack = () => {
-//     return this.props.history.push("/pantry")
-//   }
-
-//   setStateUpdateIngredientTrue = () => {
-//     this.setState({
-//       updateIngredient: true
-//     })
-//   }
-
-//   handleUpdateIngredient = () => {
-//     return (
-//       <div>
-//         {this.state.updateIngredient === true &&
-//           <div>
-//             <form id="modal-content" onSubmit={this.handleSubmit}>
-//               <label>Ingredient:</label>
-//               <input id="ingredient-name" name="ingredient_name" type="text"></input>
-//               <div id="ingredient-in-stock">
-//                 <select id="in-stock" name="in_stock">
-//                   In stock:
-//               <option value="in-stock">In stock</option>
-//                   <option value="out">Out</option>
-//                   <option value="low">Low</option>
-//                 </select>
-//               </div>
-//               <label>Notes:</label>
-//               <input id="notes" name="notes" type="text"></input>
-//               <button id="close">Hit it!</button>
-//             </form>
-
-//           </div>}
-//       </div>
-//     )
-//   }
-
-//   handleDeleteIngredient = (e) => {
-//     e.preventDefault();
-//     let ingredientId = this.props.match.params.ingredientId;
-//     const url = `${config.API_ENDPOINT}/pantry/${ingredientId}`;
-//     const authToken = TokenService.getAuthToken();
-
-//     console.log(
-//       "ingredient to be sent to server is",
-//       ingredientId
-//     );
-//     fetch(url, {
-//       method: "DELETE",
-//       headers: {
-//         "content-type": "application/json",
-//         "authorization": `Bearer ${authToken}`
-//       }
-//     })
-//       .then(res => {
-//         if (!res.ok) return res.json().then(error => Promise.reject(error));
-//       })
-//       .then(data => {
-//         console.log("delete is", data);
-//         this.props.history.push("/pantry");
-//       })
-//       .catch(error => {
-//         console.error(error);
-//       });
-//   }
-
-//   render() {
-//     console.log("this.props", this.props);
-//     return (
-
-//       <div id="individual-ingredient-view">
-//         <section id="original-ingredient-data">
-//           {/* <h2 id="update-header">Your ingredient</h2> */}
-//           <span id="current-individual-ingredient">
-//             Ingredient: {this.props.location.state.ingredient_name.toLowerCase()}
-//             <br />
-//             In stock? {this.props.location.state.in_stock}
-//             <br />
-//             Notes: {this.props.location.state.notes}
-//           </span>
-//         </section>
-//         <button id="update-ingredient-button" type="submit" onClick={() => this.setStateUpdateIngredientTrue()}>
-//           Update
-//         </button>
-//         {this.handleUpdateIngredient()}
-//         <button id="delete-ingredient-button" type="submit" onClick={(e) => this.handleDeleteIngredient(e)}>
-//           Delete
-//         </button>
-//         <button id="go-back-button" type="submit" onClick={() => this.handleGoBack()}>
-//           Go back
-//         </button>
-//       </div>
-//     )
+    this.state = {
+      isOpen: false,
+      opacity: 0
+    }
+  }
 
 
+  StyledModal = Modal.styled`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: white;
+  opacity: ${props => this.props.opacity};
+  transition: opacity ease 500ms;
+  `;
 
-//   }
-// }
+  toggleModal = (e) => {
+    this.setState({
+      isOpen: !this.state.isOpen
+    })
+  }
 
-// export default IndividualIngredient;
+  afterOpen = () => {
+    this.setState({
+      opacity: 1
+    });
+  }
+
+  beforeClose = () => {
+    this.setState({
+      opacity: 0
+    });
+  }
+
+  handleSubmit = (e) => {
+    let ingredientId = this.props.id;
+
+    const url = `${config.API_ENDPOINT}/pantry/${ingredientId}`;
+    const authToken = TokenService.getAuthToken();
+    let { ingredient_name, in_stock, notes } = e.target;
+    let updatedIngredient = {
+      id: ingredientId,
+      ingredient_name: ingredient_name ? ingredient_name.value : this.props.ingredient_name,
+      in_stock: in_stock ? in_stock.value : this.props.in_stock,
+      notes: notes ? notes.value : this.props.notes
+    };
+
+    fetch(url, {
+      method: "PATCH",
+      headers: {
+        "content-type": "application/json",
+        "authorization": `Bearer ${authToken}`
+      },
+      body: JSON.stringify(updatedIngredient)
+    })
+      .then(res => {
+        if (!res.ok) return res.json().then(error => Promise.reject(error));
+      })
+      .then(data => {
+        this.props.getIngredients();
+        this.toggleModal();
+      })
+      .catch(error => {
+        console.error(error);
+      });
+  };
+
+  handleDeleteIngredient = () => {
+    let ingredientId = this.props.id;
+    const url = `${config.API_ENDPOINT}/pantry/${ingredientId}`;
+    const authToken = TokenService.getAuthToken();
+
+    fetch(url, {
+      method: "DELETE",
+      headers: {
+        "content-type": "application/json",
+        "authorization": `Bearer ${authToken}`
+      }
+    })
+      .then(res => {
+        if (!res.ok) return res.json().then(error => Promise.reject(error));
+      })
+      .then(data => {
+        this.props.getIngredients();
+        this.toggleModal();
+      })
+      .catch(error => {
+        console.error(error);
+      });
+  }
+
+
+  fadingBackground = styled(BaseModalBackground)`
+    opacity: ${props => props.opacity};
+    transition: opacity ease 200ms;
+    `;
+
+  render() {
+    return (
+      <ModalProvider backgroundComponent={this.fadingBackground}>
+        <div className="modal-container">
+          <button className="edit-ingredient-button" onClick={this.toggleModal}>View/Edit</button>
+          <this.StyledModal
+            isOpen={this.state.isOpen}
+            afterOpen={this.afterOpen}
+            beforeClose={this.beforeClose}
+            onBackgroundClick={this.toggleModal}
+            onEscapeKeydown={this.toggleModal}
+            opacity={this.state.opacity}
+            backgroundProps={this.state.opacity}
+          >
+            <div className="styled-modal-div">
+              <p id="current-ingredient-to-edit"></p>
+              <br />
+              <form id="modal-content"
+                onSubmit={this.handleSubmit}
+              >
+                <label>Ingredient:</label>
+                <input
+                  id="ingredient-name"
+                  name="ingredient_name"
+                  type="text"
+                  placeholder={this.props.ingredient_name}></input>
+                <div id="ingredient-in-stock">
+                  <select
+                    id="in-stock"
+                    name="in_stock"
+                  >
+                    In stock:
+              <option value="in-stock">In stock</option>
+                    <option value="out-of-stock">Out</option>
+                    <option value="low">Low</option>
+                  </select>
+                </div>
+                <label>Notes:</label>
+                <input
+                  id="notes"
+                  name="notes"
+                  type="text"
+                  placeholder={this.props.notes}
+                >
+                </input>
+                <button id="update-ingredient-button" type="submit">
+                  Update
+                </button>
+              </form>
+              <button id="delete-ingredient-button" type="submit" onClick={this.handleDeleteIngredient}>
+                Delete
+              </button>
+            </div>
+          </this.StyledModal>
+        </div >
+      </ModalProvider>
+    );
+  }
+}
