+++ src/Components/Pantry/IndividualIngredient.test.js
@@ -1,20 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import DashboardRoute from './DashboardRoute';
+import IndividualIngredient from './IndividualIngredient';
 import { BrowserRouter } from 'react-router-dom';
 import renderer from "react-test-renderer";
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
   ReactDOM.render(<BrowserRouter>
-    <DashboardRoute />
+    <IndividualIngredient />
   </BrowserRouter>, div);
   ReactDOM.unmountComponentAtNode(div);
 });
 
 it('renders the UI as expected', () => {
   const tree = renderer.create(<BrowserRouter>
-    <DashboardRoute />
+    <IndividualIngredient />
   </BrowserRouter>)
     .toJSON();
   expect(tree).toMatchSnapshot();
